Batch Importing -- Warnings below!
*WARNINGS*
- database migration necessary
- gem added:  rubyist-aasm
- plugin added: delayed_job
- background process worker added: rake task

- - - -
Run these scripts after updating:
1) rake gems:install
2) rake db:migrate
3) rake jobs:work

*Strong Advice*: Test locally before deploying into production.

- - - -
This commit changes the way batch imports work.  Batch imports are now associated with a User.  A User can import works for the System or directly to a Person.  Upon upload, the works are "queued" (via acts_as_state_machine) and prepared for processing via Delayed Job (DJ).

More on DJ here:
http://railscasts.com/episodes/171-delayed-job

More on AASM here:
http://rails.aizatto.com/2007/05/24/ruby-on-rails-finite-state-machine-plugin-acts_as_state_machine/


DJ is a new plugin that streamlines background processes.  Imports can be huge (300 citations or so takes a long time to process) so they are turned over to DJ.  Once the import is complete, an email is sent to the User's email address -- in development mode, this will appear in the log/development.log, in production, you'll need to make sure your SMTP settings are correct.

When the import is complete and email sent, you'll receive a link to "review" the import.  Here you add PenNames (if importing for a Person) and "accept" or "reject" the import.  If adding works to a Person, accepting the import will auto-verify the associated Contributorships.

More documentation to appear on the google code site.

git-svn-id: https://bibapp.googlecode.com/svn/trunk@1144 6a0f952e-c525-0410-88e5-87db5789ed7d