Pen names: Add a name string
If the name string already exists, but the "name" is different
e.g., capitalized differently, comma in a different place
the "name" will be changed/overwritten with the new name.

If this approach proves not to work, we can consider having
multiple name strings with the same machine name, but 
different versions of the "name".