Alter validation: Switch the name_string_id and work_id around (since it feels more natural to me to think of looking at a work and restricting the name_strings on it than the other way around) and add role to the constraint.
Now the same name string can be added to a work in different roles (e.g. allowing the same person to be both and author and editor for a book).

Note that there is some work on the database side that needs to be done for this to work which will be in following commits.

git-svn-id: https://bibapp.googlecode.com/svn/trunk@1558 6a0f952e-c525-0410-88e5-87db5789ed7d