A small change is deleting the migration that creates the admin user. It's already in the seed, and since there are changes to the user model since that migration was written the migration actually errors out.
The main change here is dealing with the removal of model saving with
callbacks skipped. There were a lot of calls to this. It appears that
this was needed because in Rails 2 observer callbacks happened before
model callbacks. This has been fixed in Rails 3, allowing a lot of the
observers to have their code shipped back to the models and the observers
themselves to be removed.

In many cases, it also appears that an after save callback was being used
when a before save callback was more reasonable. E.g. instead of saving,
then updating a calculated field, then saving again, it makes more sense
to calculate in a before_save and let the normal save take care of saving.
So a lot of the callbacks have been changed in this way - the exceptions
are those where it looked like it was actually important to have something
happen after the save (e.g. when an id is set based on the object's id, which
will only exist after the save).

I've also removed all of the explictly defined 'after_save' (etc.) callbacks
that I found, in favor of using the class methods:

after_save :callback_name

which I think are preferential stylistically.

In the few cases where it seemed like it was actually necessary to skip
a callback in some cases this is done by making the callback conditional
(e.g. after_save :callback_name, :if => :callback_should_be_skipped), adding
an attr_accessor :callback_should_be_skipped to the class definition, and
then setting that to true before doing a save in the appropriate cases.

A few migrations needed to be modified as well to remove the
save_without_callback calls, but it doesn't seem to matter in these cases.
The migrations will still run from scratch (this is how I found the initial
migration problem mentioned at the top). If they cause a problem migrating
from an old db for someone then we'll deal with that problem at the time.