Added the prototype_legacy_helper. I'm going to back out some of the changes here and trust this to make the existing javascript compatible. I've done an example of this and it does work. It also required a bit of fiddling with the routes to make sure that the necessary routes are available.
So I'll go back and revert the changes I made to use :remote => true to the
Rails 2 form and at the same time check the routes that they use and
make sure that they are available via the routes file. I think that this will
restore the current javascript to functionality.