Extracted Strings.
I think that some more refactoring will be possible now. I need to
check where these forms/field partials are being called. If it's
only here then I think at the initial call we can globalize the
form and klass variables to avoid having to pass them. This will
clean up a lot of the parameters. And almost if not all of the labels
now come from the translations, so those will go away (if the
klass is available to generate them in place). Finally, this may make
the forms partials themselves simpler, i.e. if a lot of the render
calls are almost identical.

Before doing that I want to break up the volume/issue partial.
There doesn't seem any reason that it needs to be one, and it will
facilitate the above sort of refactoring.