Use jquery-rails gem (and its generated files) for javascript. Adjust the layout to include the relevant stuff now.
At this state the leftover prototype should be broken, but I think I
needed to make this change now to start getting it converted to jquery,
as the javascript handler for remote forms etc. will depend on the
particular framework. So as development is done there will need to be
some bouncing back and forth between this and master to see how things
work and then get them back working in this branch.