Add more caching.
Note that Rails caching is a bit funny, so there is some custom code
here to make it work better for our needs. By default it works via
controller actions, but by registering Sweepers as Observers in
application.rb one can make it work via models. Some of the stuff in
CacheHelper helps with this:
- the sweeper, in order to call expire_fragment, needs to have a
controller attached, which ensure_controller can do
- the usual fragment naming (:action => X, :controller => Y) etc.
that the cache and expire_fragment methods use require not just a
controller but a request. Since working on the model side we don't
necessarily have this, CacheHelper provides methods for caching
and expiring that convert this kind of naming into a string for use
with the provided methods. With strings the url_for machinery doesn't
get called and so this isn't a problem. These methods also provide us
an opportunity to deal with locale automatically, which we take.

This is used to cache the index page for publications. Note that this
happens at two levels - the whole page and each row of the index table.
The sweeper expires these as appropriate when a Work, Publication, or
Publisher changes.

Obviously this could be used in more places, but I want to commit this as
an example of how one might go about doing it.