Remove the PublisherSource model and db table. Remove all "uses" of it (there were no real ones).
For now I'm retaining the publisher_source_id field for publishers,
as this does get set depending on how a publisher is created.

However nothing was done with actual PublisherSource objects (we don't
even have any in our database). If we desire this kind of functionality
later it'll be easier to just build it from scratch.