Removed one use of HTMLEntities - Builder can do this itself (actually from the docs I'm surprised it's not doing it automatically. This may be because we're also using namespacing and maybe that doesn't work together with automatic escaping).
The only remaining use of HTMLEntities is in CitationParser/BaseXMLParser.
I'm not sure if we're using something else that can replace it there
(where we're decoding rather than encoding a string). Still investigating.
It's not a big deal to leave it in, but I'd rather not if it's
completely replacable with other stuff we're already using.